+++ .gitignore
@@ -19,3 +19,4 @@ examples/clingo/pydoc/clingo.html
 cmake/FindBISON.cmake
 doc/py/clingo
 doc/py/Gemfile.lock
+.mypy_cache/
