+++ app/clingo/tests/python/observer.lp
@@ -1,6 +1,6 @@
 #script (python)
 
-from clingo import Function, Tuple
+from clingo import Function, Tuple_
 
 class Observer:
     def __init__(self):
@@ -35,20 +35,20 @@ class Observer:
     def _rule(self, choice, head, body):
         head = sorted(set([ self.__map(atm) for atm in head ]))
         body = sorted(set([ self.__map(lit) for lit in body ]))
-        self.__reified.append(Function("rule", [choice, Tuple(head), Tuple(body)]))
+        self.__reified.append(Function("rule", [choice, Tuple_(head), Tuple_(body)]))
 
     def _weight_rule(self, choice, head, lower_bound, body):
         head = sorted(set([ self.__map(atm) for atm in head ]))
-        body = sorted(set([ Tuple([self.__map(lit), weight]) for lit, weight in body ]))
-        self.__reified.append(Function("weight_rule", [choice, Tuple(head), lower_bound, Tuple(body)]))
+        body = sorted(set([ Tuple_([self.__map(lit), weight]) for lit, weight in body ]))
+        self.__reified.append(Function("weight_rule", [choice, Tuple_(head), lower_bound, Tuple_(body)]))
 
     def _minimize(self, priority, literals):
-        literals = sorted(set([ Tuple([self.__map(lit), weight]) for lit, weight in literals ]))
-        self.__reified.append(Function("minimize", [priority, Tuple(literals)]))
+        literals = sorted(set([ Tuple_([self.__map(lit), weight]) for lit, weight in literals ]))
+        self.__reified.append(Function("minimize", [priority, Tuple_(literals)]))
 
     def _project(self, atoms):
         atoms = sorted(set([ self.__map(atm) for atm in atoms ]))
-        self.__reified.append(Function("project", [Tuple(atoms)]))
+        self.__reified.append(Function("project", [Tuple_(atoms)]))
 
     def output_atom(self, symbol, atom):
         self.__symbols[atom] = symbol
@@ -56,26 +56,26 @@ class Observer:
 
     def _output_term(self, symbol, condition):
         condition = sorted(set([ self.__map(lit) for lit in condition ]))
-        self.__reified.append(Function("output_term", [symbol, Tuple(condition)]))
+        self.__reified.append(Function("output_term", [symbol, Tuple_(condition)]))
 
     def _output_csp(self, symbol, value, condition):
         condition = sorted(set([ self.__map(lit) for lit in condition ]))
-        self.__reified.append(Function("output_csp", [symbol, value, Tuple(condition)]))
+        self.__reified.append(Function("output_csp", [symbol, value, Tuple_(condition)]))
 
     def _external(self, atom, value):
         self.__reified.append(Function("external", [self.__map(atom), str(value)]))
 
     def _assume(self, literals):
         literals = sorted(set([ self.__map(lit) for lit in literals ]))
-        self.__reified.append(Function("assume", [Tuple(literals)]))
+        self.__reified.append(Function("assume", [Tuple_(literals)]))
 
     def _heuristic(self, atom, type, bias, priority, condition):
         condition = sorted(set([ self.__map(lit) for lit in condition ]))
-        self.__reified.append(Function("heuristic", [self.__map(atom), str(type), bias, Tuple(condition)]))
+        self.__reified.append(Function("heuristic", [self.__map(atom), str(type), bias, Tuple_(condition)]))
 
     def _acyc_edge(self, node_u, node_v, condition):
         condition = sorted(set([ self.__map(lit) for lit in condition ]))
-        self.__reified.append(Function("acyc_edge", [node_u, node_v, Tuple(condition)]))
+        self.__reified.append(Function("acyc_edge", [node_u, node_v, Tuple_(condition)]))
 
     def theory_term_number(self, term_id, number):
         self.__terms[term_id] = lambda: number
@@ -87,13 +87,13 @@ class Observer:
         self.__terms[term_id] = lambda: Function(self.__terms[name_id_or_type]().name, [self.__terms[i]() for i in arguments])
 
     def theory_element(self, element_id, terms, condition):
-        self.__elems[element_id] = lambda: Function("elem", [Tuple([self.__terms[i]() for i in terms]), Tuple(sorted(set([ self.__map(lit) for lit in condition ])))])
+        self.__elems[element_id] = lambda: Function("elem", [Tuple_([self.__terms[i]() for i in terms]), Tuple_(sorted(set([ self.__map(lit) for lit in condition ])))])
 
     def _theory_atom(self, atom_id_or_zero, term_id, elements):
-        self.__symbols[atom_id_or_zero] = Function("theory", [self.__terms[term_id](), Tuple(sorted(set([ self.__elems[e]() for e in elements ])))]);
+        self.__symbols[atom_id_or_zero] = Function("theory", [self.__terms[term_id](), Tuple_(sorted(set([ self.__elems[e]() for e in elements ])))]);
 
     def _theory_atom_with_guard(self, atom_id_or_zero, term_id, elements, operator_id, right_hand_side_id):
-        self.__symbols[atom_id_or_zero] = Function("theory", [self.__terms[term_id](), Tuple(sorted(set([ self.__elems[e]() for e in elements ]))), self.__terms[operator_id](), self.__terms[right_hand_side_id]()]);
+        self.__symbols[atom_id_or_zero] = Function("theory", [self.__terms[term_id](), Tuple_(sorted(set([ self.__elems[e]() for e in elements ]))), self.__terms[operator_id](), self.__terms[right_hand_side_id]()]);
 
     def end_step(self):
         self.__reified.append(Function("end_step", []))
