+++ libpyclingo/clingo.pyi
@@ -1,182 +1,790 @@
-from typing import *
+# pylint: disable=invalid-name, missing-module-docstring, unused-argument, missing-function-docstring, dangerous-default-value, line-too-long
+# pylint: disable=multiple-statements, bad-whitespace, function-redefined, missing-class-docstring, unsubscriptable-object, redefined-builtin
 
-from clingo_abc import Application, Observer, Propagator
+# we can only support python 3.7 but nothing before with this
+from __future__ import annotations
 
+from typing import Any, Callable, ContextManager, Iterator, List, Mapping, MutableSequence, Optional, Sequence, Tuple, Union
+from abc import ABC, ABCMeta, abstractmethod
+
+
+class Application(ABC):
+    """
+    Interface that has to be implemented to customize clingo.
+
+    Attributes
+    ----------
+    program_name: str = 'clingo'
+        Optional program name to be used in the help output.
+
+    message_limit: int = 20
+        Maximum number of messages passed to the logger.
+    """
+
+    @abstractmethod
+    def main(self, control: Control, files: List[str]) -> None:
+        """
+        Function to replace clingo's default main function.
+
+        Parameters
+        ----------
+        control : Control
+            The main control object.
+        files : List[str]
+            The files passed to clingo_main.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def register_options(self, options: ApplicationOptions) -> None:
+        """
+        Function to register custom options.
+
+        Parameters
+        ----------
+        options : ApplicationOptions
+            Object to register additional options
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def validate_options(self) -> bool:
+        """
+        Function to validate custom options.
+
+        This function should return false or throw an exception if option
+        validation fails.
+
+        Returns
+        -------
+        bool
+        """
+
+    @abstractmethod
+    def logger(self, code: MessageCode, message: str) -> None:
+        """
+        Function to intercept messages normally printed to standard error.
+
+        By default, messages are printed to stdandard error.
+
+        Parameters
+        ----------
+        code : MessageCode
+            The message code.
+        message : str
+            The message string.
+
+        Returns
+        -------
+        None
+
+        Notes
+        -----
+        This function should not raise exceptions.
+        """
+
+
+class Observer(ABC):
+    @abstractmethod
+    def init_program(self, incremental: bool) -> None:
+        """
+        Called once in the beginning.
+
+        Parameters
+        ----------
+        incremental : bool
+            Whether the program is incremental. If the incremental flag is
+            true, there can be multiple calls to `Control.solve`.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def begin_step(self) -> None:
+        """
+        Marks the beginning of a block of directives passed to the solver.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def rule(self, choice: bool, head: List[int], body: List[int]) -> None:
+        """
+        Observe rules passed to the solver.
+
+        Parameters
+        ----------
+        choice : bool
+            Determines if the head is a choice or a disjunction.
+        head : List[int]
+            List of program atoms forming the rule head.
+        body : List[int]
+            List of program literals forming the rule body.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def weight_rule(self, choice: bool, head: List[int], lower_bound: int,
+                    body: List[Tuple[int,int]]) -> None:
+        """
+        Observe rules with one weight constraint in the body passed to the
+        solver.
+
+        Parameters
+        ----------
+        choice : bool
+            Determines if the head is a choice or a disjunction.
+        head : List[int]
+            List of program atoms forming the head of the rule.
+        lower_bound:
+            The lower bound of the weight constraint in the rule body.
+        body : List[Tuple[int,int]]
+            List of weighted literals (pairs of literal and weight) forming the
+            elements of the weight constraint.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def minimize(self, priority: int, literals: List[Tuple[int,int]]) -> None:
+        """
+        Observe minimize directives (or weak constraints) passed to the
+        solver.
+
+        Parameters
+        ----------
+        priority : int
+            The priority of the directive.
+        literals : List[Tuple[int,int]]
+            List of weighted literals whose sum to minimize (pairs of literal
+            and weight).
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def project(self, atoms: List[int]) -> None:
+        """
+        Observe projection directives passed to the solver.
+
+        Parameters
+        ----------
+        atoms : List[int]
+            The program atoms to project on.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def output_atom(self, symbol: Symbol, atom: int) -> None:
+        """
+        Observe shown atoms passed to the solver.  Facts do not have an
+        associated program atom. The value of the atom is set to zero.
+
+        Parameters
+        ----------
+        symbol : Symbolic
+            The symbolic representation of the atom.
+        atom : int
+            The associated program atom (0 for facts).
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def output_term(self, symbol: Symbol, condition: List[int]) -> None:
+        """
+        Observe shown terms passed to the solver.
+
+        Parameters
+        ----------
+        symbol : Symbol
+            The symbolic representation of the term.
+        condition : List[int]
+            List of program literals forming the condition when to show the
+            term.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def output_csp(self, symbol: Symbol, value: int,
+                   condition: List[int]) -> None:
+        """
+        Observe shown csp variables passed to the solver.
+
+        Parameters
+        ----------
+        symbol : Symbol
+            The symbolic representation of the variable.
+        value : int
+            The integer value of the variable.
+        condition : List[int]
+            List of program literals forming the condition when to show the
+            variable with its value.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def external(self, atom: int, value: TruthValue) -> None:
+        """
+        Observe external statements passed to the solver.
+
+        Parameters
+        ----------
+        atom : int
+            The external atom in form of a program literal.
+        value : TruthValue
+            The truth value of the external statement.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def assume(self, literals: List[int]) -> None:
+        """
+        Observe assumption directives passed to the solver.
+
+        Parameters
+        ----------
+        literals : List[int]
+            The program literals to assume (positive literals are true and
+            negative literals false for the next solve call).
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def heuristic(self, atom: int, type: HeuristicType, bias: int,
+                  priority: int, condition: List[int]) -> None:
+        """
+        Observe heuristic directives passed to the solver.
+
+        Parameters
+        ----------
+        atom : int
+            The program atom heuristically modified.
+        type : HeuristicType
+            The type of the modification.
+        bias : int
+            A signed integer.
+        priority : int
+            An unsigned integer.
+        condition : List[int]
+            List of program literals.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def acyc_edge(self, node_u: int, node_v: int,
+                  condition: List[int]) -> None:
+        """
+        Observe edge directives passed to the solver.
+
+        Parameters
+        ----------
+        node_u : int
+            The start vertex of the edge (in form of an integer).
+        node_v : int
+            Тhe end vertex of the edge (in form of an integer).
+        condition : List[int]
+            The list of program literals forming th condition under which to
+            add the edge.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def theory_term_number(self, term_id: int, number: int) -> None:
+        """
+        Observe numeric theory terms.
+
+        Parameters
+        ----------
+        term_id : int
+            The id of the term.
+        number : int
+            The value of the term.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def theory_term_string(self, term_id : int, name : str) -> None:
+        """
+        Observe string theory terms.
+
+        Parameters
+        ----------
+        term_id : int
+            The id of the term.
+        name : str
+            The string value of the term.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def theory_term_compound(self, term_id: int, name_id_or_type: int,
+                             arguments: List[int]) -> None:
+        """
+        Observe compound theory terms.
+
+        Parameters
+        ----------
+        term_id : int
+            The id of the term.
+        name_id_or_type : int
+            The name or type of the term where
+            - if it is -1, then it is a tuple
+            - if it is -2, then it is a set
+            - if it is -3, then it is a list
+            - otherwise, it is a function and name_id_or_type refers to the id
+            of the name (in form of a string term)
+        arguments : List[int]
+            The arguments of the term in form of a list of term ids.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def theory_element(self, element_id: int, terms: List[int],
+                       condition: List[int]) -> None:
+        """
+        Observe theory elements.
+
+        Parameters
+        ----------
+        element_id : int
+            The id of the element.
+        terms : List[int]
+            The term tuple of the element in form of a list of term ids.
+        condition : List[int]
+            The list of program literals forming the condition.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def theory_atom(self, atom_id_or_zero: int, term_id: int,
+                    elements: List[int]) -> None:
+        """
+        Observe theory atoms without guard.
+
+        Parameters
+        ----------
+        atom_id_or_zero : int
+            The id of the atom or zero for directives.
+        term_id : int
+            The term associated with the atom.
+        elements : List[int]
+            The elements of the atom in form of a list of element ids.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def theory_atom_with_guard(self, atom_id_or_zero: int, term_id: int,
+                               elements: List[int], operator_id: int,
+                               right_hand_side_id: int) -> None:
+        """
+        Observe theory atoms with guard.
+
+        Parameters
+        ----------
+        atom_id_or_zero : int
+            The id of the atom or zero for directives.
+        term_id : int
+            The term associated with the atom.
+        elements : List[int]
+            The elements of the atom in form of a list of element ids.
+        operator_id : int
+            The id of the operator (a string term).
+        right_hand_side_id : int
+            The id of the term on the right hand side of the atom.
+
+        Returns
+        -------
+        None
+        """
+
+    @abstractmethod
+    def end_step(self) -> None:
+        """
+        Marks the end of a block of directives passed to the solver.
+
+        This function is called right before solving starts.
+
+        Returns
+        -------
+        None
+        """
+
+
+class Propagator(ABC):
+    @abstractmethod
+    def init(self, init: PropagateInit) -> None:
+        """
+        This function is called once before each solving step.
+
+        It is used to map relevant program literals to solver literals, add
+        watches for solver literals, and initialize the data structures used
+        during propagation.
+
+        Parameters
+        ----------
+        init : PropagateInit
+            Object to initialize the propagator.
+
+        Returns
+        -------
+        None
+
+        Notes
+        -----
+        This is the last point to access theory atoms.  Once the search has
+        started, they are no longer accessible.
+        """
+
+    @abstractmethod
+    def propagate(self, control: PropagateControl, changes: List[int]) -> None:
+        """
+        Can be used to propagate solver literals given a partial assignment.
+
+        Parameters
+        ----------
+        control : PropagateControl
+            Object to control propagation.
+        changes : List[int]
+            List of watched solver literals assigned to true.
+
+        Returns
+        -------
+        None
+
+        Notes
+        -----
+        Called during propagation with a non-empty list of watched solver
+        literals that have been assigned to true since the last call to either
+        propagate, undo, (or the start of the search) - the change set. Only
+        watched solver literals are contained in the change set. Each literal
+        in the change set is true w.r.t. the current Assignment.
+        `PropagateControl.add_clause` can be used to add clauses. If a clause
+        is unit resulting, it can be propagated using
+        `PropagateControl.propagate`. If either of the two methods returns
+        False, the propagate function must return immediately.
+
+            c = ...
+            if not control.add_clause(c) or not control.propagate(c):
+                return
+
+        Note that this function can be called from different solving threads.
+        Each thread has its own assignment and id, which can be obtained using
+        `PropagateControl.id`.
+        """
+
+    @abstractmethod
+    def undo(self, thread_id: int, assignment: Assignment,
+             changes: List[int]) -> None:
+        """
+        Called whenever a solver with the given id undos assignments to watched
+        solver literals.
+
+        Parameters
+        ----------
+        thread_id : int
+            The solver thread id.
+        assignment : Assignment
+            Object for inspecting the partial assignment of the solver.
+        changes : List[int]
+            The list of watched solver literals whose assignment is undone.
+
+        Returns
+        -------
+        None
+
+        Notes
+        -----
+        This function is meant to update assignment dependent state in a
+        propagator but not to modify the current state of the solver.
+        """
+
+    @abstractmethod
+    def check(self, control: PropagateControl) -> None:
+        """
+        This function is similar to propagate but is called without a change
+        set on propagation fixpoints.
+
+        When exactly this function is called, can be configured using the @ref
+        PropagateInit.check_mode property.
+
+        Parameters
+        ----------
+        control : PropagateControl
+            Object to control propagation.
+
+        Returns
+        -------
+        None
+
+        Notes
+        -----
+        This function is called even if no watches have been added.
+        """
+
+    @abstractmethod
+    def decide(self, thread_id: int, assignment: Assignment, fallback: int) -> int:
+        """
+        This function allows a propagator to implement domain-specific
+        heuristics.
+
+        It is called whenever propagation reaches a fixed point.
+
+        Parameters
+        ----------
+        thread_id : int
+            The solver thread id.
+        assignment : Assignment
+            Object for inspecting the partial assignment of the solver.
+        fallback : int
+            The literal choosen by the solver's heuristic.
+
+        Returns
+        -------
+        int
+            Тhe next solver literal to make true.
+
+        Notes
+        -----
+        This function should return a free solver literal that is to be
+        assigned true. In case multiple propagators are registered, this
+        function can return 0 to let a propagator registered later make a
+        decision. If all propagators return 0, then the fallback literal is
+        used.
+        """
 
 Infimum: Symbol
 Supremum: Symbol
-
 __version__: str
 
-def Function(name: str, arguments: Iterable[Symbol] = (), positive: bool = True) -> Symbol: ...
+
+def Function(name: str, arguments: List[Symbol]=[], positive: bool=True) -> Symbol: ...
 def Number(number: int) -> Symbol: ...
 def String(string: str) -> Symbol: ...
-def Tuple(arguments: Iterable[Symbol]) -> Symbol: ...
-def clingo_main(application: Application, files: Iterable[str] = ()) -> int: ...
-def parse_program(program: str, callback: Callable[[AST], None]) -> None: ...
-def parse_term(string: str, logger: Callable[[MessageCode, str], None] = None, message_limit: int = 20) -> Symbol: ...
+def Tuple_(arguments: List[Symbol]) -> Symbol: ...
+def clingo_main(application: Application, files: List[str]=[]) -> int: ...
+def parse_program(program: str, callback: Callable[[ast.AST], None]) -> None: ...
+def parse_term(string: str, logger: Callable[[MessageCode,str],None]=None, message_limit: int=20) -> Symbol: ...
 
-class ApplicationOptions:
-    def add_flag(self, group: str, option: str, description: str, parser: Callable[[str], bool], multi: bool = False, argument: str = None) -> None: ...
+
+class ApplicationOptions(metaclass=ABCMeta):
+    def add(self, group: str, option: str, description: str, parser: Callable[[str], bool], multi: bool=False, argument: str=None) -> None: ...
     def add_flag(self, group: str, option: str, description: str, target: Flag) -> None: ...
 
-class Assignment:
-    decision_level: int
-    has_conflict: bool
-    is_total: bool
-    max_size: int
-    root_level: int
-    size: int
-
+class Assignment(Sequence[int], metaclass=ABCMeta):
     def decision(self, level: int) -> int: ...
-    def has_literal(self, literal: int) -> bool: ...
+    def has_literal(self, literal : int) -> bool: ...
     def is_false(self, literal: int) -> bool: ...
     def is_fixed(self, literal: int) -> bool: ...
     def is_true(self, literal: int) -> bool: ...
     def level(self, literal: int) -> int: ...
     def value(self, literal) -> Optional[bool]: ...
+    decision_level: int
+    has_conflict: bool
+    is_total: bool
+    root_level: int
+    trail: Trail
 
-class Backend:
-    def add_acyc_edge(self, node_u: int, node_v: int, condition: Iterable[int]) -> None: ...
-    def add_assume(self, literals: Iterable[int]) -> None: ...
-    def add_atom(self, symbol: Optional[Symbol] = None) -> int: ...
-    def add_external(self, atom: int, value: TruthValue = TruthValue.False_) -> None: ...
-    def add_heuristic(self, atom: int, type: HeuristicType, bias: int, priority: int, condition: Iterable[int]) -> None: ...
-    def add_minimize(self, priority: int, literals: Iterable[Tuple[int, int]]) -> None: ...
-    def add_project(self, atoms: Iterable[int]) -> None: ...
-    def add_rule(self, head: Iterable[int], body: Iterable[int] = (), choice: bool = False) -> None: ...
-    def add_weight_rule(self, head: Iterable[int], lower: int, body: Iterable[Tuple[int, int]], choice: bool = False) -> None: ...
+class Backend(ContextManager[Backend], metaclass=ABCMeta):
+    def add_acyc_edge(self, node_u: int, node_v: int, condition: List[int]) -> None: ...
+    def add_assume(self, literals: List[int]) -> None: ...
+    def add_atom(self, symbol : Optional[Symbol]=None) -> int: ...
+    def add_external(self, atom : int, value : TruthValue=TruthValue.False_) -> None: ...
+    def add_heuristic(self, atom: int, type: HeuristicType, bias: int, priority: int, condition: List[int]) -> None: ...
+    def add_minimize(self, priority: int, literals: List[Tuple[int,int]]) -> None: ...
+    def add_project(self, atoms: List[int]) -> None: ...
+    def add_rule(self, head: List[int], body: List[int]=[], choice: bool=False) -> None: ...
+    def add_weight_rule(self, head: List[int], lower: int, body: List[Tuple[int,int]], choice: bool=False) -> None: ...
 
-class Configuration:
-    keys: Optional[Iterable[str]]
+class Configuration(metaclass=ABCMeta):
+    def description(self, name: str) -> str: ...
+    keys: Optional[List[str]]
 
 class Control:
-    def __init__(self, arguments: Iterable[str] = (), logger: Callable[[MessageCode, str], None] = None, message_limit: int = 20): ...
-
+    def __init__(self, arguments: List[str]=[], logger: Callable[[MessageCode,str],None]=None, message_limit: int=20): ...
+    def add(self, name: str, parameters: List[str], program: str) -> None: ...
+    def assign_external(self, external: Union[Symbol,int], truth: Optional[bool]) -> None: ...
+    def backend(self) -> Backend: ...
+    def builder(self) -> ProgramBuilder: ...
+    def cleanup(self) -> None: ...
+    def get_const(self, name: str) -> Optional[Symbol]: ...
+    def ground(self, parts: List[Tuple[str,List[Symbol]]], context: Any=None) -> None: ...
+    def interrupt(self) -> None: ...
+    def load(self, path: str) -> None: ...
+    def register_observer(self, observer: Observer, replace: bool=False) -> None: ...
+    def register_propagator(self, propagator: Propagator) -> None: ...
+    def release_external(self, symbol: Union[Symbol,int]) -> None: ...
+    def solve(self, assumptions: List[Union[Tuple[Symbol,bool],int]]=[], on_model: Callable[[Model],Optional[bool]]=None, on_statistics : Callable[[StatisticsMap,StatisticsMap],None]=None, on_finish: Callable[[SolveResult],None]=None, yield_: bool=False, async_: bool=False) -> Union[SolveHandle,SolveResult]: ...
     configuration: Configuration
     is_conflicting: bool
     statistics: dict
     symbolic_atoms: SymbolicAtoms
     theory_atoms: TheoryAtomIter
-    use_enumeration_assumption: bool
-
-    def add(self, name: str, parameters: Iterable[str], program: str) -> None: ...
-    def assign_external(self, external: Union[Symbol, int], truth: Optional[bool]) -> None: ...
-    def backend(self) -> Backend: ...
-    def builder(self) -> ProgramBuilder: ...
-    def cleanup(self) -> None: ...
-    def get_const(self, name: str) -> Optional[Symbol]: ...
-    def ground(self, parts: Iterable[Tuple[str, List[Symbol]]], context: Any = None) -> None: ...
-    def interrupt(self) -> None: ...
-    def load(self, path: str) -> None: ...
-    def register_observer(self, observer: Observer, replace: bool = False) -> None: ...
-    def register_propagator(self, propagator: Propagator) -> None: ...
-    def release_external(self, symbol: Union[Symbol, int]) -> None: ...
-    def solve(self, assumptions: Iterable[Union[Tuple[Symbol, bool], int]] = (), on_model: Callable[[Model], Optional[bool]] = None, on_statistics: Callable[[StatisticsMap, StatisticsMap], None] = None, on_finish: Callable[[SolveResult], None] = None, yield_: bool = False, async_: bool = False) -> Union[SolveHandle, SolveResult]: ...
+    set_use_enumeration_assumption: bool
 
 class Flag:
-    def __init__(self, value: bool = False):
-        self.value = value
+    def __init__(self, value: bool=False): ...
+    flag: bool
 
-class HeuristicType:
+class HeuristicType(metaclass=ABCMeta):
+    Factor: HeuristicType
+    False_: HeuristicType
+    Init: HeuristicType
     Level: HeuristicType
     Sign: HeuristicType
-    Factor: HeuristicType
-    Init: HeuristicType
     True_: HeuristicType
-    False_: HeuristicType
 
-class MessageCode:
-    OperationUndefined: MessageCode
-    RuntimeError: MessageCode
+class MessageCode(metaclass=ABCMeta):
     AtomUndefined: MessageCode
     FileIncluded: MessageCode
-    VariableUnbounded: MessageCode
     GlobalVariable: MessageCode
+    OperationUndefined: MessageCode
     Other: MessageCode
+    RuntimeError: MessageCode
+    VariableUnbounded: MessageCode
 
-class Model:
+class Model(metaclass=ABCMeta):
+    def contains(self, atom: Symbol) -> bool: ...
+    def extend(self, symbols: List[Symbol]) -> None: ...
+    def is_true(self, literal: int) -> bool: ...
+    def symbols(self, atoms: bool=False, terms: bool=False, shown: bool=False, csp: bool=False, theory: bool=False, complement: bool=False) -> List[Symbol]: ...
     context: SolveControl
-    cost: Iterable[int]
+    cost: List[int]
     number: int
     optimality_proven: bool
     thread_id: int
     type: ModelType
 
-    def contains(self, atom: Symbol) -> bool: ...
-    def extend(self, symbols: Iterable[Symbol]) -> None: ...
-    def is_true(self, literal: int) -> bool: ...
-    def symbols(self, atoms: bool = False, terms: bool = False, shown: bool = False, csp: bool = False, complement: bool = False) -> Iterable[Symbol]: ...
-
-class ModelType:
+class ModelType(metaclass=ABCMeta):
     BraveConsequences: ModelType
     CautiousConsequences: ModelType
     StableModel: ModelType
 
-class ProgramBuilder:
-    def add(self, statement: AST) -> None: ...
+class ProgramBuilder(ContextManager[ProgramBuilder], metaclass=ABCMeta):
+    def add(self, statement: ast.AST) -> None: ...
 
-class PropagateControl:
-    assignment: Assignment
-    thread_id: int
-
-    def add_clause(self, clause: Iterable[int], tag: bool = False, lock: bool = False) -> bool: ...
+class PropagateControl(metaclass=ABCMeta):
+    def add_clause(self, clause: List[int], tag: bool=False, lock: bool=False) -> bool: ...
     def add_literal(self) -> int: ...
-    def add_nogood(self, clause: Iterable[int], tag: bool = False, lock: bool = False) -> bool: ...
+    def add_nogood(self, clause: List[int], tag: bool=False, lock: bool=False) -> bool: ...
     def add_watch(self, literal: int) -> None: ...
     def has_watch(self, literal: int) -> bool: ...
     def propagate(self) -> bool: ...
     def remove_watch(self, literal: int) -> None: ...
+    assignment: Assignment
+    thread_id: int
 
-class PropagateInit:
+class PropagateInit(metaclass=ABCMeta):
+    def add_clause(self, clause: List[int]) -> bool: ...
+    def add_literal(self) -> int: ...
+    def add_minimize(self, literal: int, weight: int, priority: int=0) -> None: ...
+    def add_watch(self, literal: int, thread_id: Optional[int]=None) -> None: ...
+    def add_weight_constraint(self, literal: int, literals: List[Tuple[int,int]], bound: int, type: int=0, compare_equal: bool=False) -> bool: ...
+    def propagate(self) -> bool: ...
+    def solver_literal(self, literal: int) -> int: ...
     assignment: Assignment
     check_mode: PropagatorCheckMode
     number_of_threads: int
     symbolic_atoms: SymbolicAtoms
-    theory_atoms: TheoryAtomIter
+    theory_atom: TheoryAtomIter
 
-    def add_clause(self, clause: Iterable[int]) -> None: ...
-    def add_watch(self, literal: int, thread_id: Optional[int] = None) -> None: ...
-    def solver_literal(self, literal: int) -> int: ...
-
-class PropagatorCheckMode:
+class PropagatorCheckMode(metaclass=ABCMeta):
+    Both: PropagatorCheckMode
+    Fixpoint: PropagatorCheckMode
     Off: PropagatorCheckMode
     Total: PropagatorCheckMode
-    Fixpoint: PropagatorCheckMode
 
-class SolveControl:
+class Slice(metaclass=ABCMeta):
+    pass
+
+class SolveControl(metaclass=ABCMeta):
+    def add_clause(self, literals: List[Union[Tuple[Symbol,bool],int]]) -> None: ...
+    def add_nogood(self, literals: List[Union[Tuple[Symbol,bool],int]]) -> None: ...
     symbolic_atoms: SymbolicAtoms
 
-    def add_clause(self, literals: Iterable[Union[Tuple[Symbol, bool], int]]) -> None: ...
-    def add_nogood(self, literals: Iterable[Union[Tuple[Symbol, bool], int]]) -> None: ...
-
-class SolveHandle:
+class SolveHandle(ContextManager[SolveHandle], metaclass=ABCMeta):
     def cancel(self) -> None: ...
+    def core(self) -> List[int]: ...
     def get(self) -> SolveResult: ...
     def resume(self) -> None: ...
-    def wait(self, timeout: Optional[float] = None) -> Optional[bool]: ...
+    def wait(self, timeout: Optional[float]=None) -> Optional[bool]: ...
 
-class SolveResult:
+class SolveResult(metaclass=ABCMeta):
     exhausted: bool
-    interrupted: bool
+    interruped: bool
     satisfiable: Optional[bool]
     unknown: bool
     unsatisfiable: Optional[bool]
 
-class StatisticsArray:
+class StatisticsArray(MutableSequence[Union[StatisticsArray,StatisticsMap,float]], metaclass=ABCMeta):
     def append(self, value: Any) -> None: ...
     def extend(self, values: Sequence[Any]) -> None: ...
     def update(self, values: Sequence[Any]) -> None: ...
 
-class StatisticsMap:
-    def items(self) -> Iterable[Tuple[str, Union[StatisticsArray, StatisticsMap, float]]]: ...
-    def keys(self) -> Iterable[str]: ...
-    def update(self, values: Mappping[str, Any]) -> None: ...
-    def values(self) -> Iterable[Union[StatisticsArray, StatisticsMap, float]]: ...
+class StatisticsMap(Mapping[str,Union[StatisticsArray,StatisticsMap,float]], metaclass=ABCMeta):
+    def items(self) -> List[Tuple[str, Union[StatisticsArray,StatisticsMap,float]]]: ...
+    def keys(self) -> List[str]: ...
+    def update(self, values: Mapping[str,Any]) -> None: ...
+    def values(self) -> List[Union[StatisticsArray,StatisticsMap,float]]: ...
 
-
-class Symbol:
-    arguments: Iterable[Symbol]
+class Symbol(metaclass=ABCMeta):
+    def match(self, name: str, arity: int) -> bool: ...
+    arguments: List[Symbol]
     name: str
     negative: bool
     number: int
@@ -184,61 +792,61 @@ class Symbol:
     string: str
     type: SymbolType
 
-    def match(self, name: str, arity: int) -> bool: ...
-
-class SymbolType:
-    Number: SymbolType
-    String: SymbolType
+class SymbolType(metaclass=ABCMeta):
     Function: SymbolType
     Infimum: SymbolType
+    Number: SymbolType
+    String: SymbolType
     Supremum: SymbolType
 
-class SymbolicAtom:
+class SymbolicAtom(metaclass=ABCMeta):
+    def match(self, name: str, arity: int) -> bool: ...
     is_external: bool
     is_fact: bool
     literal: int
     symbol: Symbol
 
-    def match(self, name: str, arity: int) -> bool: ...
-
-class SymbolicAtomIter(Iterator[SymbolicAtom]):
+class SymbolicAtomIter(Iterator[SymbolicAtom], metaclass=ABCMeta):
     pass
 
-class SymbolicAtoms:
-    signatures: Iterable[Tuple[str, int, bool]]
+class SymbolicAtoms(Mapping[Union[Symbol,int],SymbolicAtom], metaclass=ABCMeta):
+    def by_signature(self, name: str, arity: int, positive: bool=True) -> Iterator[Symbol]: ...
+    signatures: List[Tuple[str,int,bool]]
 
-    def by_signature(self, name: str, arity: int, positive: bool = True) -> Iterator[Symbol]: ...
-
-class TheoryAtom:
-    elements: Iterable[TheoryElement]
+class TheoryAtom(metaclass=ABCMeta):
+    elements: List[TheoryElement]
     guard: Tuple[str, TheoryTerm]
     literal: int
     term: TheoryTerm
 
-class TheoryAtomIter(Iterator[SymbolicAtom]):
+class TheoryAtomIter(Iterator[TheoryAtom], metaclass=ABCMeta):
     pass
 
-class TheoryElement:
-    condition: Iterable[TheoryTerm]
+class TheoryElement(metaclass=ABCMeta):
+    condition: List[TheoryTerm]
     condition_id: int
-    terms: Iterable[TheoryTerm]
+    terms: List[TheoryTerm]
 
-class TheoryTerm:
-    arguments: Iterable[Symbol]
+class TheoryTerm(metaclass=ABCMeta):
+    arguments: List[TheoryTerm]
     name: str
     number: int
     type: TheoryTermType
 
-class TheoryTermType:
+class TheoryTermType(metaclass=ABCMeta):
     Function: TheoryTermType
-    Number: TheoryTermType
-    Symbol: TheoryTermType
     List: TheoryTermType
-    Tuple: TheoryTermType
+    Number: TheoryTermType
     Set: TheoryTermType
+    Symbol: TheoryTermType
+    Tuple: TheoryTermType
 
-class TruthValue:
-    True_: TruthValue
+class Trail(Sequence[int], metaclass=ABCMeta):
+    def begin(self, level: int) -> int: ...
+    def end(self, level: int) -> int: ...
+
+class TruthValue(metaclass=ABCMeta):
     False_: TruthValue
     Free: TruthValue
     Release: TruthValue
+    True_: TruthValue
